+++ ethereum/src/constants/errors.ts
